requirements.txt
requests==2.31.0
streamlit==1.32.0
selenium==4.18.1
webdriver_manager==4.0.1
markdownify==0.11.6
xhtml2pdf==0.2.15
python-docx==1.1.0
htmldocx==0.0.6
beautifulsoup4==4.12.3
